backend/pyproject.toml
[project]
name = "crm-protegrt-backend"
version = "0.1.0"
description = "CRM Seguros Protegrt - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "geoalchemy2>=0.15.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "argon2-cffi>=23.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "openpyxl>=3.1.5",
    "python-socketio>=5.12.0",
    "jinja2>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
